+++ source/blender/nodes/NOD_geometry_nodes_eval_log.hh
@@ -171,7 +171,7 @@ struct NodeWithWarning {
 
 struct NodeWithExecutionTime {
   DNode node;
-  int exec_time;
+  uint64_t exec_time;
 };
 
 /** The same value can be referenced by multiple sockets when they are linked. */
@@ -207,7 +207,7 @@ class LocalGeoLogger {
   void log_value_for_sockets(Span<DSocket> sockets, GPointer value);
   void log_multi_value_socket(DSocket socket, Span<GPointer> values);
   void log_node_warning(DNode node, NodeWarningType type, std::string message);
-  void log_execution_time(DNode node, int exec_time);
+  void log_execution_time(DNode node, uint64_t exec_time);
 };
 
 /** The root logger class. */
@@ -281,7 +281,7 @@ class NodeLog {
   Vector<SocketLog> input_logs_;
   Vector<SocketLog> output_logs_;
   Vector<NodeWarning, 0> warnings_;
-  int exec_time_ = 0;
+  uint64_t exec_time_ = 0;
 
   friend ModifierLog;
